Refactor code — Add "updateFile" to utils
