+++ lib/automations/release/utils/index.js
@@ -5,4 +5,5 @@ module.exports = {
 	getReleaseBranch: require( './get-release-branch' ),
 	duplicateChecker: require( './duplicate-pr-checker' ),
 	hasMilestone: require( './has-milestone' ),
+	updateFile: require( './update-file' ),
 };
